third commit move main class to the package
