+++ src/main/java/spring_app/mongodb/Config.java
@@ -1,4 +1,4 @@
-package mongodb;
+package spring_app.mongodb;
 
 import com.mongodb.ConnectionString;
 import com.mongodb.MongoClientSettings;
