+++ src/main/java/spring_app/user/User.java
@@ -1,4 +1,4 @@
-package user;
+package spring_app.user;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.index.Indexed;
