+++ src/main/java/spring_app/user/repository/UserRepository.java
@@ -1,7 +1,7 @@
-package user.repository;
+package spring_app.user.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import user.User;
+import spring_app.user.User;
 
 public interface UserRepository extends MongoRepository<User, Long> {
     User findUserByUsername(String username);
