+++ src/main/resources/application.properties
@@ -5,4 +5,5 @@ spring.data.mongodb.database=usersDatabase
 
 #logging
 logging.level.org.springframework.data=debug
+spring.main.allow-bean-definition-overriding=true
